requirements.txt
streamlit
pymupdf
pytesseract
Pillow
langchain
